Configure img capture template & python settings
The capture template is an attempt to make an "easy" way to insert the
standard image info into org templates. I'm not sure that it works.

The Python config is a failed attempt to be able to use elpy to run
unittests using the M-x compile & M-x recompile. However, there's an
issue where the tests need to run from the root directory of the
repository and not the test subdirectory. I did some hacking on this but
it doesn't work.